requirements.txt
pdfplumber
sentence-transformers
faiss-cpu
transformers
torch
langchain
python-dotenv
numpy
pytest
streamlit
